WIP: T11 -- Split app-detail.test.tsx
