+++ design/specs/016-handler-telemetry-debt/tasks/T11-split-app-detail-tests.md
@@ -1,7 +1,7 @@
 ---
 task_id: "T11"
 title: "Split app-detail.test.tsx"
-status: "planned"
+status: "done"
 depends_on: ["T10"]
 implements: ["FR#11", "AC#3"]
 ---
